Second round of code review fixes
