+++ lib/phobos.rb
@@ -22,7 +22,7 @@ require 'phobos/errors'
 require 'phobos/listener'
 require 'phobos/actions/process_batch'
 require 'phobos/actions/process_message'
-require 'phobos/actions/process_message_batch'
+require 'phobos/actions/process_batch_inline'
 require 'phobos/producer'
 require 'phobos/handler'
 require 'phobos/batch_handler'
